requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Babel==4.0.0
python-dotenv==1.0.1
gunicorn==22.0.0
email-validator==2.1.1
Pillow==10.3.0
